// apps/pokeverse/src/helpers/types.ts
export interface PokemonSprite {
    front_default: string;
    front_shiny: string;
    back_default?: string;
    back_shiny?: string;
    other: {
        'official-artwork': {
            front_default: string;
            front_shiny: string;
        };
        home: {
            front_default: string;
            front_shiny: string;
        };
        showdown: {
            front_default: string;
            back_default: string;
        };
    };
}

export interface PokemonDetail {
    id: number;
    name: string;
    base_experience: number;
    height: number;
    weight: number;
    types: Array<{
        type: { name: string };
    }>;
    abilities: Array<{
        ability: { name: string; url: string };
        is_hidden: boolean;
    }>;
    sprites: PokemonSprite;
    stats: Array<{
        base_stat: number;
        stat: { name: string };
    }>;
    moves: Array<{
        move: { name: string; url: string };
    }>;
    species: {
        name: string;
        url: string;
    };
}
// ...
export interface MoveApiResponse {
    id: number;
    name: string;
    type: { name: string };
    power: number | null;
    accuracy: number | null;
    damage_class: { name: string };
}
// ...
export interface BattleMove {
    id: number;
    name: string;
    type: string;
    power: number;
    accuracy: number;
    category: 'Physical' | 'Special';
}

export interface BattlePokemon {
    id: number;
    name: string;
    types: string[];
    level: number;
    currentHP: number;
    maxHP: number;
    selectedMoves: BattleMove[];
    backSprite: string;
    frontSprite: string;
    calculatedStats: {
        hp: number;
        attack: number;
        defense: number;
        specialAttack: number;
        specialDefense: number;
        speed: number;
    };
}